pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hospital_mgmt_env"
version = "0.1.0"
description = "Hospital Mgmt Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "streamlit>=1.20.0",
    "pandas>=1.5.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m hospital_mgmt_env.server.app
server = "hospital_mgmt_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["hospital_mgmt_env", "hospital_mgmt_env.server"]
package-dir = { "hospital_mgmt_env" = ".", "hospital_mgmt_env.server" = "server" }